lib: fix panic from 0xFF opt subneg in truncated msg
Iff the parser state decides 0xFF is a locally supported and enabled
enabled option, and a truncated subnegotiation for opt code 0xFF is
received, don't panic from a negative range end.

Previously in `TelnetParser::process` when handling
a `EventType::SubNegotiation` buffer the processing code would look for
a buffer ending in `IAC SE` and then try to pull out remaining data
using a range expression like `&buffer[3..len - 2]`. If the buffer
happens to be 4 bytes long, and the negotiated option was itself 0xFF,
and it ended with a single `SE` byte,  then the check for the proper
ending will pass, thinking the option code is the IAC indicator for the
SE command. This will result in the range expression being `[3..2]`,
which will panic.

The fix applied in this commit is to require that the buffer end in IAC
SE _and_ that `len - 2` is greater or equal to 3.

Discovered via the cargo fuzzer integration.